+++ scripts/get_senna.sh
@@ -7,6 +7,6 @@ wget https://github.com/baojie/senna/raw/master/embeddings/embeddings.txt
 wget https://github.com/baojie/senna/raw/master/hash/words.lst
 cd ..
 echo "Processing data"
-python -c "from shlo.utils import senna; senna()"
+python -c "from shlo.data_utils import senna; senna()"
 rm data/embeddings.txt
 rm data/words.lst
